Bump Rpc pcap sizes from int32 to int64
This commit bumps the types of all Rpc Pcap*Size fields from int32 to int64. Though it would be crazy, theoretically someone may want to retrieve PCAP greater than ~2.14GB.

+++ config/config.go
@@ -54,9 +54,9 @@ type RpcConfig struct {
         ServerCert              string
         Port                    int
         PcapPath                string
-        PcapLimitSize           int32
-        PcapClientChunkSize     int32
-        PcapClientMaxSize       int32
+        PcapLimitSize           int64
+        PcapClientChunkSize     int64
+        PcapClientMaxSize       int64
 }
 
 // Config is a json-decoded configuration for running stenographer.
